Cleaned up access and fixed a bug with w/ssyntax.
The method access, had a redundent case if the second argument was a
symbol. This case is already handled by the default case. Secondly,
w/ssyntax had incorrectly assumed that ssyntax-macros and
ssyntax-sym-macs were defined for all of the symbols in syms. This
problem was fixed by changing map to trues, so that if one of them
isn't defined, it isn't included in the macrolet or symbol-macrolet binding.