Added iterate to Clamp.
I've been playing with it recently, but ran into some issues when I
tried to use both Clamp and iterate in the same package. I decided
that the easiest thing to do would be to replace any symbol in the
iter form with the iterate equivalent if it exists (ie clamp:while
will be replaced with iter:while).