Added back iterate.
What I decided to do was to have the conflicting macros try to
determine if there are in an iterate clause. If they are, they should
expand into the iterate form. I did this for all of the conflicting
macros except for summing and in. Instead of using the iterate version
of summing, just use sum. As for in I still have to figure out what to
do since there is no easy way to determine if it is in an iterate clause.